requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
cachetools==5.5.2
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
Flask==3.1.2
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.38.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
itsdangerous==2.2.0
mkl-service==2.4.0
packaging==25.0
platformdirs==4.4.0
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
regex==2025.9.1
rsa==4.9.1
safetensors==0.6.2
sentry-sdk==2.38.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.22.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
